% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{create_amendment}
\alias{create_amendment}
\title{Function to generate an amendment data table for unit testing}
\usage{
create_amendment()
}
\value{
data table with amendment information
}
\description{
Function to generate an amendment data table for unit testing
}
