% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{create_parms}
\alias{create_parms}
\title{Function to generate a parameter data table for unit testing}
\usage{
create_parms()
}
\value{
data table with values for RothC parameters
}
\description{
Function to generate a parameter data table for unit testing
}
