% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{create_rotation}
\alias{create_rotation}
\title{Function to generate a data table with rotation information for unit testing}
\usage{
create_rotation()
}
\value{
data table with rotation information
}
\description{
Function to generate a data table with rotation information for unit testing
}
