% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{create_soil_properties}
\alias{create_soil_properties}
\title{Function to generate a soil properties data table for unit testing}
\usage{
create_soil_properties()
}
\value{
data table with soil properties information
}
\description{
Function to generate a soil properties data table for unit testing
}
