% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{create_weather}
\alias{create_weather}
\title{Function to generate a weather data table for unit testing}
\usage{
create_weather()
}
\value{
data table with weather information
}
\description{
Function to generate a weather data table for unit testing
}
