% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_calculate_bd}
\alias{rc_calculate_bd}
\title{Function to calculate the dry soil bulk density based on Dutch pedotransfer functions}
\usage{
rc_calculate_bd(dt)
}
\arguments{
\item{dt}{(data table) Contains the columns A_SOM_LOI (organic matter content, \%) and A_CLAY_MI (clay content \%)}
}
\value{
Data table with the dry soil bulk density (g/cm)
}
\description{
Function to calculate the dry soil bulk density based on Dutch pedotransfer functions
}
