% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_check_inputs}
\alias{rc_check_inputs}
\title{Function to check input tables of soil, crop, and amendment data}
\usage{
rc_check_inputs(soil_properties, rothc_rotation, rothc_amendment)
}
\arguments{
\item{soil_properties}{(list) List with soil properties: A_C_OF, soil organic carbon content (g/kg) or B_C_ST03, soil organic carbon stock (Mg C/ha), preferably for soil depth 0.3 m; A_CLAY_MI, clay content (\%); A_DENSITY_SA, dry soil bulk density (g/cm3)}

\item{rothc_rotation}{(data.table) Table with crop rotation details and crop management actions that have been taken. Includes also crop inputs for carbon. See details for desired format.}

\item{rothc_amendment}{(data.table) A table with the following column names: P_DATE_FERTILIZATION, P_ID, P_NAME, P_DOSE, P_C_OF, P_C_OF_INPUT, and P_HC.}
}
\value{
Error messages indicating if input data is not in order
}
\description{
Function to check input tables of soil, crop, and amendment data
}
