% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_extend_amendments}
\alias{rc_extend_amendments}
\title{Function to extend the user amendments input file to cover the full range of simulation years}
\usage{
rc_extend_amendments(amendments, start_date, end_date = NULL, simyears = NULL)
}
\arguments{
\item{amendments}{(data table) Input amendments table for the RothC model. See details for further information}

\item{start_date}{(character, formatted YYYY-MM-DD) Date in which simulation starts}

\item{end_date}{(character, formatted YYYY-MM-DD) Required if simyears is not supplied}

\item{simyears}{(numeric) number of simulation years of the RothC model}
}
\value{
An extended amendment input file to be used in the rotsee package
}
\description{
Function to extend the user amendments input file to cover the full range of simulation years
}
\details{
amendments: amendment table
Includes the columns:
* P_ID (character), ID of the soil amendment product
* P_NAME (character), name of the soil amendment product, optional
* P_C_OF_INPUT (numeric), the organic carbon input from soil amendment product on a field level (kg C/ha)
* P_DOSE (numeric), applied dose of soil amendment product (kg/ha), required if P_C_OF_INPUT is not supplied
* P_C_OF (numeric), organic carbon content of the soil amendment product (g C/kg), required if P_C_OF_INPUT is not supplied
* P_HC (numeric), the humification coefficient of the soil amendment product (fraction)
* P_DATE_FERTILIZATION (date), date of fertilizer application (formatted YYYY-MM-DD)
}
