% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_extend_crops}
\alias{rc_extend_crops}
\title{Function to extend the user crop input file to cover the full range of simulation years}
\usage{
rc_extend_crops(simyears, crops, start_date)
}
\arguments{
\item{simyears}{(numeric) number of simulation years of the RothC model}

\item{crops}{(data table) Input crop table for the RothC model. See details for further information}

\item{start_date}{(formatted YYYY-MM-DD) Date in which simulation starts}
}
\value{
An extended crop input file to be used in the main RothC model
}
\description{
Function to extend the user crop input file to cover the full range of simulation years
}
\details{
Crops: crop table
Includes the columns: 
* B_LU_START (start of crop rotation),
* B_LU_END (end of crop rotation),
* B_LU (a crop id), 
* B_LU_NAME (a crop name, optional),
* B_LU_HC, the humification coefficient of crop organic matter (-). When not supplied, default RothC value will be used
* B_C_OF_INPUT, the organic carbon input on field level (kg C/ha)
}
