% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_initialise.R
\name{rc_initialise}
\alias{rc_initialise}
\title{Function for initializing the RothC model}
\usage{
rc_initialise(B_LU_BRP, A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{A_CLAY_MI}{(numeric) value for the clay content of the soil}
}
\description{
Function for initializing the RothC model
}
