% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_input_crop.R
\name{rc_input_crop}
\alias{rc_input_crop}
\title{Do check and expand crop input table for RothC for the Netherlands}
\usage{
rc_input_crop(dt, cf_yield = 1)
}
\arguments{
\item{dt}{(data.table) Table with crop rotation and related crop properties for Carbon input. See details for information.}

\item{cf_yield}{(numeric) A yield correction factor (fraction) if yield is higher than regional average}
}
\description{
Helper function to check the content and format of the crop input table.
}
\details{
The crop table used as input for carbon modelling requires at minimum data on effective organic matter inputs and related year.
To run this function, the dt should contain the following columns:
* year
* month
* B_LU (a crop id)
* B_LU_NAME (a crop name, optional):
* B_LU_HC (the humification coefficient of crop organic matter (-). When not supplied, default RothC value will be used)
* B_C_OF_INPUT (the organic carbon input on field level (kg C/ha). In case not known, can be calculated using function \link{rc_calculate_B_C_OF})
}
