% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_input_events.R
\name{rc_input_events}
\alias{rc_input_events}
\title{Combine all EVENT data needed for RothC modelling}
\usage{
rc_input_events(crops, amendment)
}
\arguments{
\item{crops}{(data.table) Table containing carbon inputs of DPM and RPM from crops across the entire simulation period, as calculated in \link{rc_input_event_crop}}

\item{amendment}{(data.table) Table containing carbon inputs of DPM, RPM, and HUM from amendments across the entire simulation period, as calculated in \link{rc_input_event_amendment}}
}
\description{
This function combines required inputs into a data.table that is needed as input for the RothC model.
}
