% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_input_rmf.R
\name{rc_input_rmf}
\alias{rc_input_rmf}
\title{Estimate the rate modifying factors for the RothC modelling}
\usage{
rc_input_rmf(dt = NULL, B_DEPTH = 0.3, A_CLAY_MI, dt.weather, dt.time)
}
\arguments{
\item{dt}{(data.table) Table with crop rotation and related crop properties for Carbon input.}

\item{B_DEPTH}{(numeric) Depth of the cultivated soil layer (m), simulation depth. Default set to 0.3.}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{dt.weather}{(data.table) Data table of monthly weather}

\item{dt.time}{(data.table) table with all combinations of year and month in the simulation period}
}
\description{
This function prepares the rate modifying factors for RothC given crop cover
}
\details{
dt: crop rotation table
contains at least the following columns:
* B_LU_START (date), start of crop growth
* B_LU_END (DATE), end of crop growth

dt.weather: weather table
contains the following columns:
* year (optional)
* month
* W_TEMP_MEAN_MONTH
* W_PREC_SUM_MONTH
* W_ET_POT_MONTH
* W_ET_ACT_MONTH
}
