% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_maxval}
\alias{rc_maxval}
\title{Get maximum value of provided parameter from the pandex data table}
\usage{
rc_maxval(this.parameter)
}
\arguments{
\item{this.parameter}{(character) parameter name for which a maximum value is needed}
}
\value{
maximum allowed value of the supplied parameter
}
\description{
Get maximum value of provided parameter from the pandex data table
}
