% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_minval}
\alias{rc_minval}
\title{Get minimum value of provided parameter from the pandex data table}
\usage{
rc_minval(this.parameter)
}
\arguments{
\item{this.parameter}{(character) parameter name for which a minimum value is needed}
}
\value{
minimum allowed value of the supplied parameter
}
\description{
Get minimum value of provided parameter from the pandex data table
}
