% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_ode.R
\name{rc_ode}
\alias{rc_ode}
\title{RothC ordinary differential equation}
\usage{
rc_ode(time, y, parms)
}
\arguments{
\item{time}{(numeric) time to print output}

\item{y}{state variables to run RothC}

\item{parms}{(list) parameters to run RothC, includes rate modifying functions and constants R1, abc, and d}
}
\description{
Set of differential equations to calculate the evolution of Soil Organic Carbon stocks using the RothC model.
}
