% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_multicore.R
\name{rc_parallel}
\alias{rc_parallel}
\title{RothC parallel function}
\usage{
rc_parallel(this.xs, dt.c, scen, simyears = 50, p = NULL, final = TRUE)
}
\arguments{
\item{this.xs}{(numeric) selected id for a single field}

\item{dt.c}{(data.table) set will all fields}

\item{scen}{(character) scenarios to be simulated. Options include BAU, BAUIMPR, CLT and ALL.}

\item{simyears}{(integer) value for the amount of years to simulate, default is 50 years}

\item{p}{(progress bar) progress bar}

\item{final}{(boolean) option to select only the last year}
}
\description{
Function to run RothC parallel for a series of fields
}
