% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_parstable}
\alias{rc_parstable}
\title{function to read in and save table with relevant data on used parameters in the rotsee package}
\usage{
rc_parstable()
}
\value{
package table with relevant information of parameters used in rotsee
}
\description{
function to read in and save table with relevant data on used parameters in the rotsee package
}
