% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_sim_multi.R
\name{rc_sim_multi}
\alias{rc_sim_multi}
\title{Parallel RothC calculations}
\usage{
rc_sim_multi(
  soil_properties,
  rotation,
  amendment,
  A_DEPTH = 0.3,
  B_DEPTH = 0.3,
  parms,
  weather,
  final = FALSE,
  quiet = TRUE,
  strategy = "sequential",
  cores = NULL
)
}
\arguments{
\item{soil_properties}{(data.table) Data table with soil properties. For inputs, see \link{rc_sim} with additional ID column used to identify scenario.}

\item{rotation}{(data.table) Table with crop rotation details and crop management actions that have been taken. For inputs, see \link{rc_sim} with additional ID column used to identify scenario.}

\item{amendment}{(data.table) Table with amendment input details. For inputs, see \link{rc_sim} with additional ID column used to identify scenario.}

\item{A_DEPTH}{(numeric) Depth for which soil sample is taken (m). Default set to 0.3.}

\item{B_DEPTH}{(numeric) Depth of the cultivated soil layer (m), simulation depth. Default set to 0.3.}

\item{parms}{(list) A list with simulation parameters controlling the dynamics of RothC Model. For inputs, see \link{rc_sim}.}

\item{weather}{(data.table) Table with weather information. For inputs, see \link{rc_sim}.}

\item{final}{(boolean) option to select only the average of the last 10 years}

\item{quiet}{(boolean) showing progress bar for calculation RothC C-saturation for each field}

\item{strategy}{(character) select strategy for multistep calculations. Options are 'sequential' (single step calculation), 'multicore' (linux) and 'multisession' (windows)}

\item{cores}{(numeric) number of cores that can be used for multi-step calculations.}
}
\value{
A data.table containing RothC simulation results for all scenarios, with columns including the ID column from inputs plus simulation outputs from \link{rc_sim}. If final=TRUE, returns only the average of the last 10 years of simulation for each scenario.
}
\description{
Function to perform RothC calculation for multiple scenarios via \link{rc_sim} with optional parallel processing. Input parameters are similar to \link{rc_sim}, with data.table inputs requiring an additional ID column to identify each scenario.
}
