% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_sim_multi.R
\name{rc_sim_multistep}
\alias{rc_sim_multistep}
\title{RothC parallel function}
\usage{
rc_sim_multistep(
  this.xs,
  soil_properties,
  rotation = NA_real_,
  amendment = NA_real_,
  A_DEPTH = 0.3,
  B_DEPTH = 0.3,
  parms = NA_real_,
  weather = NA_real_,
  p,
  final = FALSE
)
}
\arguments{
\item{this.xs}{(numeric) selected id for a single field}

\item{soil_properties}{(data.table) Data table with soil properties. For inputs, see \link{rc_sim} with additional column 'xs' as identifier for separate scenario's}

\item{rotation}{(data.table) Table with crop rotation details and crop management actions that have been taken. For inputs, see \link{rc_sim} with additional column 'xs' as identifier for separate scenario's}

\item{amendment}{(data.table) Table with amendment input details. For inputs, see \link{rc_sim}}

\item{A_DEPTH}{(numeric) Depth for which soil sample is taken (m). Default set to 0.3.}

\item{B_DEPTH}{(numeric) Depth of the cultivated soil layer (m), simulation depth. Default set to 0.3.}

\item{parms}{(list) A list with simulation parameters controlling the dynamics of RothC Model. For inputs, see \link{rc_sim}.}

\item{weather}{(data.table) Table with weather information. For inputs, see \link{rc_sim}.}

\item{p}{(progress bar) progress bar}

\item{final}{(boolean) option to select only the average of the last 10 years}
}
\description{
Intermediary function called by \link{rc_sim_multi} to perform parallel rothc calculations
}
