% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_time_period}
\alias{rc_time_period}
\title{Function to create data table with dates and months of the entire simulation period}
\usage{
rc_time_period(start_date, end_date)
}
\arguments{
\item{start_date}{(date) start date of the simulation period (formatted as YYYY-MM-DD)}

\item{end_date}{(date) end date of the simulation period (formatted as YYYY-MM-DD)}
}
\value{
Table with all year and month combinations of the simulation period
}
\description{
Function to create data table with dates and months of the entire simulation period
}
