% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_visualize_plot}
\alias{rc_visualize_plot}
\title{Function to plot information on C pools when rc_sim is run in visualize mode}
\usage{
rc_visualize_plot(dt, event = NULL, save_dir = getwd())
}
\arguments{
\item{dt}{(data.table) data table with monthly information on the state of the pools CDPM, CRPM, CBIO, and CHUM}

\item{event}{(data.table) data table with information on c inputs, contains columns time, var, and value}

\item{save_dir}{(character) directory to save PNGs; defaults to current working directory}
}
\value{
data table with monthly totals and changes in different C pools, plots of their trends
}
\description{
Function to plot information on C pools when rc_sim is run in visualize mode
}
