% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_calculate_B_C_OF}
\alias{rc_calculate_B_C_OF}
\title{Function to calculate the C input of your crop}
\usage{
rc_calculate_B_C_OF(dt)
}
\arguments{
\item{dt}{(data table) Table with crop rotation details and crop management actions that have been taken. For more information see details.}
}
\value{
Data table with total C input from crops
}
\description{
Function to calculate the C input of your crop
}
\details{
Crop data table
Contains the following columns:
* B_LU_YIELD (numeric), the mean crop yield (kg dry matter/ha)
* B_LU_HI (numeric), the harvest index of the crop
* B_LU_HI_RES (numeric), fraction of biomass that is residue
* B_LU_RS_FR (numeric), Root-to-shoot ratio of the crop
* M_CROPRESIDUE (logical), indicator of whether crop residue is incorporated into the soil
}
