% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_calculate_bd}
\alias{rc_calculate_bd}
\title{Function to calculate the dry soil bulk density based on Dutch pedotransfer functions}
\usage{
rc_calculate_bd(dt)
}
\arguments{
\item{dt}{(data table) Contains the columns A_CLAY_MI (clay content \%) and at least one of A_SOM_LOI (organic matter content, \%) and A_C_OF (organic carbon content, g C/kg)}
}
\value{
Data table with provided data and the calculated dry soil bulk density (g/cm3)
}
\description{
Function to calculate the dry soil bulk density based on Dutch pedotransfer functions
}
\examples{
## calculate the dry soil bulk density (g/cm3) using organic matter content (\%)
dt_om <- data.table::data.table(A_CLAY_MI = 12, # data table with legal soil information
A_SOM_LOI = 3)

rc_calculate_bd(dt_om)

## calculate the dry soil bulk density using organic carbon content (g C/kg)
dt_c <- data.table::data.table(A_CLAY_MI = 12, # data table with legal soil information
A_C_OF = 80)

rc_calculate_bd(dt_c)

}
