% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_extend_crops}
\alias{rc_extend_crops}
\title{Function to extend the user crop input file to cover the full range of simulation years}
\usage{
rc_extend_crops(crops, start_date, end_date = NULL, simyears = NULL)
}
\arguments{
\item{crops}{(data.table) Input crop table for the RothC model. See details for further information}

\item{start_date}{(character/date, formatted YYYY-MM-DD) Date in which simulation starts}

\item{end_date}{(character/date, formatted YYYY-MM-DD) Required if simyears is not supplied}

\item{simyears}{(numeric) number of simulation years of the RothC model}
}
\value{
An extended crop input file to be used in the rotsee package
}
\description{
Function to extend the user crop input file to cover the full range of simulation years
}
\details{
Crop table includes the columns:
\itemize{
\item B_LU_START (start of crop rotation),
\item B_LU_END (end of crop rotation),
\item B_LU (a crop id) optional,
\item B_LU_NAME (a crop name, optional),
\item B_LU_HC, the humification coefficient of crop organic matter (-). When not supplied, default RothC value will be used
\item B_C_OF_INPUT, the organic carbon input on field level (kg C/ha)
}
}
\examples{
## examples to extend valid crop dataset
# create valid crop set
crops <- data.table::data.table(
B_LU_START = c("2020-01-01", "2020-06-01"),
B_LU_END = c("2020-03-31", "2020-08-31"),
B_LU_HC = c(0.5, 0.3),
B_C_OF_INPUT = c(100, 200)
)

# using start_date and end_date
rc_extend_crops(crops = crops, start_date = "2020-01-01", end_date = "2022-12-31")

# using start_date and simyears
rc_extend_crops(crops = crops, start_date = "2020-01-01", simyears = 3)

}
