% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_initialise.R
\name{rc_initialise}
\alias{rc_initialise}
\title{Function for initializing the RothC model}
\usage{
rc_initialise(B_LU_BRP, soil_properties)
}
\arguments{
\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{soil_properties}{(list) List with soil properties: A_C_OF, soil organic carbon content (g/kg) or B_C_ST03, soil organic carbon stock (Mg C/ha), preferably for soil depth 0.3 m; A_CLAY_MI, clay content (\%); A_DENSITY_SA, dry soil bulk density (g/cm3)}
}
\description{
Function for initializing the RothC model
}
