% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_input_amendment.R
\name{rc_input_amendment}
\alias{rc_input_amendment}
\title{Do check and expand amendment input table for RothC}
\usage{
rc_input_amendment(dt = NULL)
}
\arguments{
\item{dt}{(data.table) Table with amendments and amendment properties for Carbon input.}
}
\description{
Helper function to check the content and format of the amendment input table.
}
\details{
The amendments table used as input for carbon modelling requires at minimum data on effective organic matter inputs and related year.
This helper function assists the checking and controlling of amendments properties involved.

rothc_amendment: amendment table
Includes the columns:
* P_ID (character), ID of the soil amendment product
* P_NAME (character), name of the soil amendment product, optional
* B_C_OF_INPUT (numeric), the organic carbon input from soil amendment product on a field level (kg C/ha)
* P_DOSE (numeric), applied dose of soil amendment product (kg/ha), required if B_C_OF_INPUT is not supplied
* P_C_OF (numeric), organic carbon content of the soil amendment product (g C/kg), required if B_C_OF_INPUT is not supplied
* P_HC (numeric), the humification coefficient of the soil amendment product (fraction)
* P_DATE_FERTILIZATION (date), date of fertilizer application (formatted YYYY-MM-DD)
}
