% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_input_event_amendment.R
\name{rc_input_event_amendment}
\alias{rc_input_event_amendment}
\title{Calculate the monthly timing of carbon inputs for different fertilizer strategies}
\usage{
rc_input_event_amendment(amendment = NULL, dt.time)
}
\arguments{
\item{amendment}{(data.table) A table with the following column names: P_ID (optional), P_NAME (optional), year, month, cin_tot, cin_hum, cin_dpm, and cin_rpm.}

\item{dt.time}{(data.table) Table containing all combinations of months and years in the simulation period, with columns year, month, and time}
}
\description{
This function calculates the timing of carbon inputs (kg C per ha) based on type of amendment application.
}
\details{
This function increases temporal detail for time series of C inputs of organic amendments.
The inputs for organic amendments are organised in the data.table amendment, where the carbon inputs has the unit kg C / ha.
dt. time Must contain integerish columns \code{year} (>= 0) and \code{month} (1..12), numeric column \code{time}, 
and unique \code{(year, month)} pairs over the simulation window.

The output is an EVENT object.
}
