% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_input_events.R
\name{rc_input_events}
\alias{rc_input_events}
\title{Combine all EVENT data needed for RothC modelling}
\usage{
rc_input_events(crops, amendment, dt.time)
}
\arguments{
\item{crops}{(data.table) Table with crop rotation, cultivation management, year and potential Carbon inputs.}

\item{amendment}{(data.table) A table with the following column names: year, month, cin_tot, cin_hum, cin_dpm, and cin_rpm.}

\item{dt.time}{(data.table) Table containing all combinations of months and years in the simulation period}
}
\description{
This function combines required inputs into a data.table that is needed as input for the RothC model.
}
