% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_input_events.R
\name{rc_input_events}
\alias{rc_input_events}
\title{Combine all EVENT data needed for RothC modelling}
\usage{
rc_input_events(crops = NULL, amendment = NULL, simyears)
}
\arguments{
\item{crops}{(data.table) Table with crop rotation, cultivation management, year and potential Carbon inputs.}

\item{amendment}{(data.table) A table with the following column names: year, month, cin_hum, cin_dpm, and cin_rpm, cin_tot (optional).}

\item{simyears}{(numeric) Amount of years for which the simulation should run, default: 50 years}
}
\description{
This function combines required inputs into a data.table that is needed as input for the RothC model.
}
