% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_multicore.R
\name{rc_multicore}
\alias{rc_multicore}
\title{Multicore RothC simulation}
\usage{
rc_multicore(
  soil_properties,
  rotation,
  amendment,
  A_DEPTH = 0.3,
  B_DEPTH = 0.3,
  parms,
  weather,
  final = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{soil_properties}{(data.table) Data table with soil properties. For inputs, see \link{rc_sim} with additional ID column used to identify scenario.}

\item{rotation}{(data.table) Table with crop rotation details and crop management actions that have been taken. For inputs, see \link{rc_sim} with additional ID column used to identify scenario.}

\item{amendment}{(data.table) Table with amendment input details. For inputs, see \link{rc_sim} with additional ID column used to identify scenario.}

\item{A_DEPTH}{(numeric) Depth for which soil sample is taken (m). Default set to 0.3.}

\item{B_DEPTH}{(numeric) Depth of the cultivated soil layer (m), simulation depth. Default set to 0.3.}

\item{parms}{(list) A list with simulation parameters controlling the dynamics of RothC Model. For inputs, see \link{rc_sim}.}

\item{weather}{(data.table) Table with weather information. For inputs, see \link{rc_sim}.}

\item{final}{(boolean) option to select only the average of the last 10 years}

\item{quiet}{(boolean) showing progress bar for calculation RothC C-saturation for each field}
}
\description{
Function to perform parallel RothC calculations via \link{rc_sim} using multicore processing.
}
