% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_set_refact}
\alias{rc_set_refact}
\title{Function to determine W_ET_REFACT given a crop profile and supplied weather table, based on Dutch Makkink factors}
\usage{
rc_set_refact(weather, crop, dt.time)
}
\arguments{
\item{weather}{(data.table) weather data table containing columns year, month, W_TEMP_MEAN_MONTH, W_PREC_SUM_MONTH, W_ET_REF_MONTH, W_ET_ACT_MONTH, W_ET_REFACT.}

\item{crop}{(data.table) data table with crop information. Contains at least the columns B_LU_START, B_LU_END, and D_MAKKINK_JAN through DEC}

\item{dt.time}{(data.table) Data table with the combination of years and months spanning the entire simulation period}
}
\value{
Weather table with W_ET_REFACT column added
}
\description{
Function to determine W_ET_REFACT given a crop profile and supplied weather table, based on Dutch Makkink factors
}
