% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_sim.R
\name{rc_sim}
\alias{rc_sim}
\title{Simulate SOC evolution using Roth-C}
\usage{
rc_sim(
  soil_properties,
  A_DEPTH = 0.3,
  B_DEPTH = 0.3,
  cf_yield = 1,
  M_TILLAGE_SYSTEM = "CT",
  rothc_rotation,
  rothc_amendment = NULL,
  rothc_parms = NULL,
  weather = NULL
)
}
\arguments{
\item{soil_properties}{(list) List with soil properties: A_C_OF, soil organic carbon content (g/kg) or B_C_ST03, soil organic carbon stock (Mg C/ha), preferably for soil depth 0.3 m; A_CLAY_MI, clay content (\%); A_DENSITY_SA, dry soil bulk density (g/cm3)}

\item{A_DEPTH}{(numeric) Depth for which soil sample is taken (m). Default set to 0.3.}

\item{B_DEPTH}{(numeric) Depth of the cultivated soil layer (m), simulation depth. Default set to 0.3.}

\item{cf_yield}{(numeric) A relative yield correction factor (fraction) if yield is higher than regional average}

\item{M_TILLAGE_SYSTEM}{(character) gives the tillage system applied. Options include NT (no-till), ST (shallow-till), CT (conventional-till) and DT (deep-till).}

\item{rothc_rotation}{(data.table) Table with crop rotation details and crop management actions that have been taken. Includes also crop inputs for carbon. See details for desired format.}

\item{rothc_amendment}{(data.table) A table with the following column names: year, month, P_NAME, P_DOSE, P_HC, P_OM, and p_p2o5, where month is optional.}

\item{rothc_parms}{(list) A list with simulation parameters controlling the dynamics of RothC Model. For more information, see details.}

\item{weather}{(data.table) Table with following column names: month, W_TEMP_MEAN_MONTH, W_PREC_MEAN_MONTH, W_ET_POT_MONTH, W_ET_ACT_MONTH. For more information, see details.}
}
\description{
This function calculates the change in carbon stock or C pools (in kg C per ha) based on organic matter amendments, crop rotation, and long-term averaged weather conditions.
}
\details{
This function simulates the fate of SOC given the impact of soil properties, weather and management.
The following inputs are mandatory: rothc_rotation, A_SOM_LOI (\%), and A_CLAY_MI (\%). All other data is optional.
When no weather inputs are given, these are estimated from long-term average weather conditions in the Netherlands.

rothc_amendment: amendment table
P_NAME is the fertilizer name, P_DOSE has units (kg / ha), P_HC is the humification coefficient (fraction), P_OM is the organic matter content (%) and p_p2o5 is the phosphate content (%)

rothc_rotation: crop table
Includes the columns: 
May additionally include the columns M_GREEN_TIMING, M_CROPRESIDUE, M_IRRIGATION and M_RENEWAL, all in upper case.
* M_GREEN_TIMING (character) the month in which the catch crop is sown, options: (august,september,october,november,never)
* M_CROPRESIDUE (boolean) whether crop residues are amended to the soil after harvest.
* M_IRRIGATION (boolean) whether the crop is irrigated.
* M_RENEWAL (boolean) whether the grassland is renewed (only applicable for grassland)

rothc_parms: simulation parameters
Table with possible parameters to adapt the calculations. May include the columns initialize, c_fractions, dec_rates, simyears and unit.
* initialize: scenario to initialize the carbon pools. Options TRUE or FALSE, default is FALSE
* c_fractions: Distribution over the different C pools.
* dec_rates: Decomposition rates of the different pools
* simyears: Duration of simulation (years), default is 50
* unit: Unit in which the output should be given. Options: 'all','soc', 'psomperfraction'

weather: Average weather conditions
Table containing columns month, W_TEMP_MEAN_MONTH (temperature in °C), W_PREC_MEAN_MONTH (precipitation in mm), W_ET_POT_MONTH (potential evapotranspiration in mm), and W_ET_ACT_MONTH. (actual evapotranspiration in mm).
If no table is given, average Dutch conditions are used
}
