% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_sim.R
\name{rc_sim}
\alias{rc_sim}
\title{Simulate SOC evolution using Roth-C}
\usage{
rc_sim(
  soil_properties,
  A_DEPTH = 0.3,
  B_DEPTH = 0.3,
  rothc_rotation = NULL,
  rothc_amendment = NULL,
  rothc_parms = NULL,
  weather = NULL,
  irrigation = NULL,
  visualize = FALSE
)
}
\arguments{
\item{soil_properties}{(data.table) Data table with soil properties. See details below.}

\item{A_DEPTH}{(numeric) Depth for which soil sample is taken (m). Default set to 0.3.}

\item{B_DEPTH}{(numeric) Depth of the cultivated soil layer (m), simulation depth. Default set to 0.3.}

\item{rothc_rotation}{(data.table) Table with crop rotation details and crop management actions that have been taken. See details below.}

\item{rothc_amendment}{(data.table) A table with amendment application details. See details below.}

\item{rothc_parms}{(list) A list with simulation parameters controlling the dynamics of RothC Model. See details below.}

\item{weather}{(data.table) Table with weather data. See details below.}

\item{irrigation}{(data.table) Table with irrigation data. See details below.}

\item{visualize}{(boolean) If TRUE, run rc_sim in visualize mode. Results are directly plotted in c pools per month to allow for direct interpretation.}
}
\value{
Table with development of C pools in the field over the given simulation period
}
\description{
This function calculates the change in carbon stock or C pools (in kg C per ha) based on organic matter amendments, crop rotation, and long-term averaged weather conditions.
}
\section{soil_properties}{

Data table with soil properties information.
\itemize{
\item A_C_OF (numeric), soil organic carbon content (g C/kg), preferably for soil depth 0.3 m
\item B_C_ST03 (numeric), soil organic carbon stock (Mg C/ha), preferably for soil depth 0.3 m. Required if A_C_OF is not supplied
\item A_CLAY_MI (numeric), clay fraction (\\%)
\item A_DENSITY_SA (numeric), dry soil bulk density(g/cm3). In case this is not know, can be calculated using function \link{rc_calculate_bd} given a clay and organic matter content
}
}

\section{rothc_amendment}{

Table with amendment application details. Can be extended to encompass the entire simulation period using \link{rc_extend_amendments}
\itemize{
\item B_C_OF_INPUT (numeric), the organic carbon input from soil amendment product on a field level (kg C/ha)
\item P_DOSE (numeric), applied dose of soil amendment product (kg/ha), required if B_C_OF_INPUT is not supplied
\item P_C_OF (numeric), organic carbon content of the soil amendment product (g C/kg), required if B_C_OF_INPUT is not supplied
\item P_HC (numeric), the humification coefficient of the soil amendment product (fraction).
\item P_DATE_FERTILIZATION (date), date of fertilizer application (formatted YYYY-MM-DD)
}
}

\section{rothc_rotation}{

Table with crop rotation details and management actions. Can be extended to encompass the entire simulation period using \link{rc_extend_crops}.
\itemize{
\item B_LU_START (start of crop rotation)
\item B_LU_END (end of crop rotation)
\item B_LU_HC, the humification coefficient of crop organic matter (-). When not supplied, default RothC value will be used
\item B_C_OF_INPUT, the organic carbon input on field level (kg C/ha). In case not known, can be calculated using function \link{rc_calculate_bcof}
}
}

\section{rothc_parms}{

List with simulation parameters controlling the RothC model calculations
\itemize{
\item initialize (boolean) scenario to initialize the carbon pools. Options TRUE or FALSE, default is FALSE.
\item c_fractions (list) Distribution over the different C pools. If not supplied nor calculated via model initialization, default RothC distribution is used
\item dec_rates (list) list of decomposition rates of the different pools. If not supplied, default RothC values are used
\item unit (character) Unit in which the output should be given. Options: 'A_SOM_LOI' (\\% organic matter),'psoc' (g C/kg), 'psomperfraction' (\\% organic matter of each fraction), 'cstock' (kg C/ha of each fraction)
\item method (character) method to solve ordinary differential equations, see \link[deSolve]{ode} for options. Default is 'adams'.
\item start_date (character, formatted "YYYY-MM-DD") Start date of simulation period. If not provided, first date of crop rotation or amendment application is taken.
\item end_date (character, formatted "YYYY-MM-DD") End date of simulation period. If not provided, last date of crop rotation or amendment application is taken.
}
}

\section{weather}{

Data table with weather information
\itemize{
\item year (integer) optional, should span the entire simulation period. If not supplied, month must include all 12 months which will be auto-expanded across simulation period
\item month
\item W_TEMP_MEAN_MONTH (temperature in °C)
\item W_PREC_SUM_MONTH (precipitation in mm)
\item W_ET_REF_MONTH (reference evapotranspiration in mm)
\item W_ET_ACT_MONTH (actual evapotranspiration in mm)
\item W_ET_REFACT (factor to recalculate reference to actual evapotranspiration, default 0.75)
}
}

\section{irrigation}{

Data table with irrigation information
\itemize{
\item B_DATE_IRRIGATION (date, formatted YYYY-MM-DD) Date of field irrigation
\item B_IRR_AMOUNT (numeric) Irrigation amount (mm)
}
}

