% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_update_parms}
\alias{rc_update_parms}
\title{Function to check user given RothC simulation parameters, or provide default if none are given}
\usage{
rc_update_parms(parms, M_TILLAGE_SYSTEM)
}
\arguments{
\item{parms}{(list) List containing the columns dec_rates, c_fractions, initialise, simyears, and unit}

\item{M_TILLAGE_SYSTEM}{}
}
\description{
Function to check user given RothC simulation parameters, or provide default if none are given
}
