% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_helpers.R
\name{rc_update_weather}
\alias{rc_update_weather}
\title{Function to check weather table and derive default table when not given as input}
\usage{
rc_update_weather(dt)
}
\arguments{
\item{dt}{(Data table) Table with following column names: month, temp, prec, et_pot, et_act.}
}
\value{
A data table containing monthly data of temperature, precipitation, and evapotranspiration.
}
\description{
Function to check weather table and derive default table when not given as input
}
